refactor: move types and interfaces to a single folder
