+++ src/components/CryptoFormSearch/index.tsx
@@ -6,18 +6,10 @@ import Select from '../Select';
 import { waitTwoMinutes } from '../../Helpers';
 import currencies from '../../services/currencies.json';
 
+import { CryptoFormSearchProps } from '../../@types';
+
 import searchIcon from '../../assets/search.svg';
 import './styles.css';
-
-interface CryptoFormSearchProps {
-  setSearch: React.Dispatch<React.SetStateAction<string>>;
-  setMarket: React.Dispatch<React.SetStateAction<string>>;
-  setSeries: React.Dispatch<React.SetStateAction<string>>;
-  search: string;
-  market: string;
-  series: string;
-  searchBySymbol: Function;
-}
 
 const CryptoFormSearch: React.FC<CryptoFormSearchProps> = (
   { setSearch,
