+++ src/components/Select/index.tsx
@@ -1,14 +1,9 @@
-import React, { SelectHTMLAttributes } from 'react';
+import React from 'react';
+import { SelectProps } from '../../@types';
 
 import CustomArrow from '../CustomArrow';
 
 import './styles.css';
-
-interface SelectProps extends SelectHTMLAttributes<HTMLSelectElement> {
-  name: string;
-  label?: string;
-  options: { label: string; value: string }[];
-}
 
 const Select: React.FC<SelectProps> = ({ name, label, options, ...rest }) => {
   return (
